pyproject.toml
[project]
name = "netpicker-cli"
version = "0.1.1"
description = "NetPicker CLI for devices and backups"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Netpicker Team", email = "support@netpicker.io" }]
dependencies = [
  "typer>=0.12.5",
  "httpx>=0.27",
  "pydantic>=2.8",
  "keyring>=25.2.1",
  "python-dotenv>=1.0.1",
  "tabulate>=0.9.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
netpicker = "netpicker_cli.cli:app"
netpicker-mcp = "netpicker_cli.mcp.server:main"
netpicker-api = "netpicker_cli.api_server:main"

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "respx>=0.21.1",
  "pytest-cov>=5",
  "ruff>=0.6",
  "black>=24",
  "mypy>=1.11",
  "types-tabulate>=0.9.0.20240106",
]
mcp = [
  "mcp>=1.0.0",
]
api = [
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.8",
]
ai = [
  "httpx>=0.27",
  "pydantic>=2.8",
]
all = [
  "mcp>=1.0.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0",
  "httpx>=0.27",
  "pydantic>=2.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
addopts = "-q --cov=netpicker_cli --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
strict = false
